🚸 improve snippet for the `storage_var` decorator
Allowing the developer to jump into the argument block even if he doesn't
need to, seems better than not offering this feature.
With this improvement, the way we write `storage_var` that support
an argument (mapping) is smoother without impacting the experience when you don't
need it. In this case, just skip the argument block using tab.